pyproject.toml
[tool.poetry]
name = "transit-data-analytics-demo"
version = "23.6.5"
description = ""
authors = ["Andrew Vaccaro <atvaccaro@gmail.com>"]
readme = "README.md"
packages = [{include = "transit_data_analytics_demo"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
huey = "^2.4.5"
redis = "^4.6.0"
google-cloud-storage = "^2.10.0"
pydantic = "^1.10.9"
backoff = "^2.2.1"
prometheus-client = "^0.17.0"
schedule = "^1.2.0"
pyyaml = "^6.0"
pendulum = "^2.1.2"
typer = "^0.9.0"
humanize = "^4.7.0"
python-dotenv = "^1.0.0"
gtfs-realtime-bindings = "^1.0.0"
tqdm = "^4.65.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.4.1"
ruff = "^0.0.275"
types-requests = "^2.31.0.1"
types-pyyaml = "^6.0.12.10"
pytest = "^7.4.0"
pytest-spec = "^3.2.0"
polyfactory = "^2.4.0"
pytest-dotenv = "^0.5.2"

[tool.black]
line-length = 120

[tool.mypy]
plugins = "pydantic.mypy"

[tool.ruff]
line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
